@@ -45,53 +48,54 @@
 		"@discordjs/ui": "workspace:^",
 		"@microsoft/api-extractor-model": "7.24.0",
 		"@microsoft/tsdoc": "0.14.1",
-		"@vercel/og": "^0.0.19",
+		"@react-icons/all-files": "^4.1.0",
+		"@vercel/og": "^0.0.21",
 		"@vscode/codicons": "^0.0.32",
 		"ariakit": "^2.0.0-next.41",
 		"cmdk": "^0.1.20",
-		"meilisearch": "^0.28.0",
-		"next": "^12.3.1",
-		"next-mdx-remote": "^4.1.0",
-		"next-progress": "^2.2.0",
-		"next-themes": "^0.2.1",
+		"meilisearch": "^0.30.0",
+		"next": "^13.0.6",
+		"next-mdx-remote": "^4.2.0",
+		"next-themes": "npm:@wits/next-themes@latest",
 		"react": "^18.2.0",
 		"react-custom-scrollbars-2": "^4.5.0",
 		"react-dom": "^18.2.0",
-		"react-icons": "^4.6.0",
 		"react-syntax-highlighter": "^15.5.0",
 		"react-use": "^17.4.0",
-		"rehype-ignore": "^1.0.1",
-		"rehype-pretty-code": "^0.4.0",
+		"rehype-ignore": "^1.0.3",
+		"rehype-pretty-code": "^0.5.1",
 		"rehype-raw": "^6.1.1",
-		"rehype-slug": "^5.0.1",
+		"rehype-slug": "^5.1.0",
 		"remark-gfm": "^3.0.1",
-		"sharp": "^0.31.1",
+		"sharp": "^0.31.2",
 		"shiki": "^0.11.1",
 		"swr": "^1.3.0"
 	},
 	"devDependencies": {
+		"@next/bundle-analyzer": "^13.0.6",
 		"@testing-library/react": "^13.4.0",
 		"@testing-library/user-event": "^14.4.3",
-		"@types/node": "16.11.68",
-		"@types/react": "^18.0.21",
-		"@types/react-dom": "^18.0.6",
+		"@types/node": "16.18.4",
+		"@types/react": "^18.0.25",
+		"@types/react-dom": "^18.0.9",
 		"@types/react-syntax-highlighter": "^15.5.5",
-		"@unocss/cli": "^0.45.30",
-		"@unocss/reset": "^0.45.30",
-		"@vitejs/plugin-react": "^2.1.0",
-		"@vitest/coverage-c8": "^0.24.3",
-		"concurrently": "^7.4.0",
+		"@unocss/cli": "^0.47.5",
+		"@unocss/reset": "^0.47.5",
+		"@vitejs/plugin-react": "^2.2.0",
+		"@vitest/coverage-c8": "^0.25.3",
+		"concurrently": "^7.6.0",
+		"cpy-cli": "^4.2.0",
 		"cross-env": "^7.0.3",
-		"eslint": "^8.25.0",
-		"eslint-config-neon": "^0.1.39",
+		"eslint": "^8.28.0",
+		"eslint-config-neon": "^0.1.40",
 		"eslint-formatter-pretty": "^4.1.0",
-		"happy-dom": "^7.6.0",
-		"prettier": "^2.7.1",
-		"prettier-plugin-tailwindcss": "^0.1.13",
-		"typescript": "^4.8.4",
-		"unocss": "^0.45.30",
-		"vercel": "^28.4.12",
-		"vitest": "^0.24.3"
+		"happy-dom": "^7.7.0",
+		"prettier": "^2.8.0",
+		"prettier-plugin-tailwindcss": "^0.2.0",
+		"typescript": "^4.9.3",
+		"unocss": "^0.47.5",
+		"vercel": "^28.7.0",
+		"vitest": "^0.25.3"
 	},
 	"engines": {
 		"node": ">=16.9.0"
