@@ -5,12 +5,20 @@
 	"private": true,
 	"scripts": {
 		"build": "turbo run build",
+		"build:affected": "turbo run build --filter='...[origin/main]'",
+		"build:apps": "turbo run build:local --filter='...{apps/*}'",
+		"build:apps:affected": "turbo run build:local --filter='...{apps/*}[origin/main]'",
 		"test": "turbo run test --parallel",
+		"test:affected": "turbo run test --filter='...[origin/main]' --parallel",
 		"lint": "turbo run lint --parallel",
+		"lint:affected": "turbo run lint --filter='...[origin/main]' --parallel",
 		"format": "turbo run format --parallel",
+		"format:affected": "turbo run format --filter='...[origin/main]' --parallel",
 		"fmt": "turbo run format --parallel",
-		"postinstall": "is-ci || husky install",
+		"fmt:affected": "turbo run format --filter='...[origin/main]' --parallel",
 		"docs": "turbo run docs --parallel",
+		"docs:affected": "turbo run docs --filter='...[origin/main]' --parallel",
+		"postinstall": "is-ci || husky install",
 		"update": "yarn upgrade-interactive",
 		"create-package": "node packages/scripts/src/packageScript.mjs"
 	},
@@ -38,19 +46,19 @@
 	},
 	"homepage": "https://discord.js.org",
 	"devDependencies": {
-		"@commitlint/cli": "^17.1.2",
-		"@commitlint/config-angular": "^17.1.0",
-		"@favware/cliff-jumper": "^1.8.8",
-		"@favware/npm-deprecate": "^1.0.5",
+		"@commitlint/cli": "^17.3.0",
+		"@commitlint/config-angular": "^17.3.0",
+		"@favware/cliff-jumper": "^1.9.0",
+		"@favware/npm-deprecate": "^1.0.6",
 		"conventional-changelog-cli": "^2.2.2",
-		"husky": "^8.0.1",
+		"husky": "^8.0.2",
 		"is-ci": "^3.0.1",
-		"lint-staged": "^13.0.3",
-		"tsup": "^6.3.0",
-		"turbo": "^1.5.6",
-		"typescript": "^4.8.4",
-		"unocss": "^0.45.30",
-		"vitest": "^0.24.3"
+		"lint-staged": "^13.0.4",
+		"tsup": "^6.5.0",
+		"turbo": "^1.6.3",
+		"typescript": "^4.9.3",
+		"unocss": "^0.47.5",
+		"vitest": "^0.25.3"
 	},
 	"resolutions": {
 		"@microsoft/tsdoc-config": "patch:@microsoft/tsdoc-config@npm:0.16.1#.yarn/patches/@microsoft-tsdoc-config-npm-0.16.1-81031b1bbf.patch"
