@@ -35,14 +35,14 @@
 		"@microsoft/tsdoc": "0.14.1"
 	},
 	"devDependencies": {
-		"@types/node": "16.11.68",
+		"@types/node": "16.18.4",
 		"cross-env": "^7.0.3",
-		"eslint": "^8.25.0",
-		"eslint-config-neon": "^0.1.39",
+		"eslint": "^8.28.0",
+		"eslint-config-neon": "^0.1.40",
 		"eslint-formatter-pretty": "^4.1.0",
-		"prettier": "^2.7.1",
-		"tsup": "^6.3.0",
-		"typescript": "^4.8.4"
+		"prettier": "^2.8.0",
+		"tsup": "^6.5.0",
+		"typescript": "^4.9.3"
 	},
 	"engines": {
 		"node": ">=16.9.0"
