@@ -26,7 +26,8 @@
 		"test": "__tests__"
 	},
 	"files": [
-		"dist"
+		"dist",
+		"scripts"
 	],
 	"contributors": [
 		"Crawl <icrawltogo@gmail.com>",
@@ -57,21 +58,21 @@
 	"dependencies": {
 		"@msgpack/msgpack": "^2.8.0",
 		"@vladfrangu/async_event_emitter": "^2.1.2",
-		"ioredis": "^5.2.3"
+		"ioredis": "^5.2.4"
 	},
 	"devDependencies": {
-		"@favware/cliff-jumper": "^1.8.8",
-		"@microsoft/api-extractor": "^7.33.4",
-		"@types/node": "^16.11.68",
-		"@vitest/coverage-c8": "^0.24.3",
+		"@favware/cliff-jumper": "^1.9.0",
+		"@microsoft/api-extractor": "^7.33.6",
+		"@types/node": "16.18.4",
+		"@vitest/coverage-c8": "^0.25.3",
 		"cross-env": "^7.0.3",
-		"eslint": "^8.25.0",
-		"eslint-config-neon": "^0.1.39",
+		"eslint": "^8.28.0",
+		"eslint-config-neon": "^0.1.40",
 		"eslint-formatter-pretty": "^4.1.0",
-		"prettier": "^2.7.1",
-		"tsup": "^6.3.0",
-		"typescript": "^4.8.4",
-		"vitest": "^0.24.3"
+		"prettier": "^2.8.0",
+		"tsup": "^6.5.0",
+		"typescript": "^4.9.3",
+		"vitest": "^0.25.3"
 	},
 	"engines": {
 		"node": ">=16.9.0"
