@@ -1,1 +1,8 @@
-module.exports = require('../../.prettierrc.json');
+module.exports = {
+	...require('../../.prettierrc.json'),
+	plugins: [
+		'prettier-plugin-astro',
+		'prettier-plugin-tailwindcss', // MUST come last
+	],
+	pluginSearchDirs: false,
+};
