@@ -1,3 +1,5 @@
+'use client';
+
 import type { TokenDocumentation } from '@discordjs/api-extractor-utils';
 import Link from 'next/link';
 
@@ -7,10 +9,12 @@ export function HyperlinkedText({ tokens }: { tokens: TokenDocumentation[] }) {
 			{tokens.map((token, idx) => {
 				if (token.path) {
 					return (
-						<Link href={token.path} key={idx} prefetch={false}>
-							<a className="text-blurple focus:ring-width-2 focus:ring-blurple rounded outline-0 focus:ring">
-								{token.text}
-							</a>
+						<Link
+							className="text-blurple focus:ring-width-2 focus:ring-blurple rounded outline-0 focus:ring"
+							href={token.path}
+							key={idx}
+						>
+							{token.text}
 						</Link>
 					);
 				}
