@@ -1,7 +1,10 @@
+'use client';
+
 import type { ApiFunctionJSON } from '@discordjs/api-extractor-utils';
-import { Menu, MenuButton, MenuItem, useMenuState } from 'ariakit';
+import { VscChevronDown } from '@react-icons/all-files/vsc/VscChevronDown';
+import { VscVersions } from '@react-icons/all-files/vsc/VscVersions';
+import { Menu, MenuButton, MenuItem, useMenuState } from 'ariakit/menu';
 import { useState } from 'react';
-import { VscChevronDown, VscVersions } from 'react-icons/vsc';
 import { DocContainer } from '../DocContainer';
 import { ParametersSection } from '../Sections';
 
