@@ -1,4 +1,15 @@
-export const PACKAGES = ['brokers', 'builders', 'collection', 'proxy', 'rest', 'util', 'voice', 'ws'];
+export const PACKAGES = [
+	'brokers',
+	'builders',
+	'collection',
+	'core',
+	'formatters',
+	'proxy',
+	'rest',
+	'util',
+	'voice',
+	'ws',
+];
 
 export const DESCRIPTION =
 	"discord.js is a powerful node.js module that allows you to interact with the Discord API very easily. It takes a much more object-oriented approach than most other JS Discord libraries, making your bot's code significantly tidier and easier to comprehend.";
