@@ -1,6 +1,7 @@
 export * as EmbedAssertions from './messages/embed/Assertions.js';
 export * from './messages/embed/Embed.js';
-export * from './messages/formatters.js';
+// TODO: Consider removing this dep in the next major version
+export * from '@discordjs/formatters';
 
 export * as ComponentAssertions from './components/Assertions.js';
 export * from './components/ActionRow.js';
@@ -11,8 +12,27 @@ export * from './components/textInput/TextInput.js';
 export * as TextInputAssertions from './components/textInput/Assertions.js';
 export * from './interactions/modals/Modal.js';
 export * as ModalAssertions from './interactions/modals/Assertions.js';
-export * from './components/selectMenu/SelectMenu.js';
-export * from './components/selectMenu/SelectMenuOption.js';
+
+export * from './components/selectMenu/BaseSelectMenu.js';
+export * from './components/selectMenu/ChannelSelectMenu.js';
+export * from './components/selectMenu/MentionableSelectMenu.js';
+export * from './components/selectMenu/RoleSelectMenu.js';
+export * from './components/selectMenu/StringSelectMenu.js';
+// TODO: Remove those aliases in v2
+export {
+	/**
+	 * @deprecated Will be removed in the next major version, use {@link StringSelectMenuBuilder} instead.
+	 */
+	StringSelectMenuBuilder as SelectMenuBuilder,
+} from './components/selectMenu/StringSelectMenu.js';
+export {
+	/**
+	 * @deprecated Will be removed in the next major version, use {@link StringSelectMenuOptionBuilder} instead.
+	 */
+	StringSelectMenuOptionBuilder as SelectMenuOptionBuilder,
+} from './components/selectMenu/StringSelectMenuOption.js';
+export * from './components/selectMenu/StringSelectMenuOption.js';
+export * from './components/selectMenu/UserSelectMenu.js';
 
 export * as SlashCommandAssertions from './interactions/slashCommands/Assertions.js';
 export * from './interactions/slashCommands/SlashCommandBuilder.js';
