@@ -1,78 +1,26 @@
 'use strict';
 
-const { SelectMenuBuilder: BuildersSelectMenu, isJSONEncodable, normalizeArray } = require('@discordjs/builders');
-const { toSnakeCase } = require('../util/Transformers');
-const { resolvePartialEmoji } = require('../util/Util');
+const process = require('node:process');
+const StringSelectMenuBuilder = require('./StringSelectMenuBuilder');
+
+let deprecationEmitted = false;
 
 /**
- * Class used to build select menu components to be sent through the API
- * @extends {BuildersSelectMenu}
+ * @deprecated Use {@link StringSelectMenuBuilder} instead.
+ * @extends {StringSelectMenuBuilder}
  */
-class SelectMenuBuilder extends BuildersSelectMenu {
-  constructor({ options, ...data } = {}) {
-    super(
-      toSnakeCase({
-        ...data,
-        options: options?.map(({ emoji, ...option }) => ({
-          ...option,
-          emoji: emoji && typeof emoji === 'string' ? resolvePartialEmoji(emoji) : emoji,
-        })),
-      }),
-    );
-  }
+class SelectMenuBuilder extends StringSelectMenuBuilder {
+  constructor(...params) {
+    super(...params);
 
-  /**
-   * Normalizes a select menu option emoji
-   * @param {SelectMenuOptionData|JSONEncodable<APISelectMenuOption>} selectMenuOption The option to normalize
-   * @returns {Array<SelectMenuOptionBuilder|APISelectMenuOption>}
-   * @private
-   */
-  static normalizeEmoji(selectMenuOption) {
-    if (isJSONEncodable(selectMenuOption)) {
-      return selectMenuOption;
+    if (!deprecationEmitted) {
+      process.emitWarning(
+        'The SelectMenuBuilder class is deprecated, use StringSelectMenuBuilder instead.',
+        'DeprecationWarning',
+      );
+      deprecationEmitted = true;
     }
-
-    const { emoji, ...option } = selectMenuOption;
-    return {
-      ...option,
-      emoji: typeof emoji === 'string' ? resolvePartialEmoji(emoji) : emoji,
-    };
-  }
-
-  /**
-   * Adds options to this select menu
-   * @param {RestOrArray<APISelectMenuOption>} options The options to add to this select menu
-   * @returns {SelectMenuBuilder}
-   */
-  addOptions(...options) {
-    return super.addOptions(normalizeArray(options).map(option => SelectMenuBuilder.normalizeEmoji(option)));
-  }
-
-  /**
-   * Sets the options on this select menu
-   * @param {RestOrArray<APISelectMenuOption>} options The options to set on this select menu
-   * @returns {SelectMenuBuilder}
-   */
-  setOptions(...options) {
-    return super.setOptions(normalizeArray(options).map(option => SelectMenuBuilder.normalizeEmoji(option)));
-  }
-
-  /**
-   * Creates a new select menu builder from json data
-   * @param {JSONEncodable<APISelectMenuComponent> | APISelectMenuComponent} other The other data
-   * @returns {SelectMenuBuilder}
-   */
-  static from(other) {
-    if (isJSONEncodable(other)) {
-      return new this(other.toJSON());
-    }
-    return new this(other);
   }
 }
 
 module.exports = SelectMenuBuilder;
-
-/**
- * @external BuildersSelectMenu
- * @see {@link https://discord.js.org/#/docs/builders/main/class/SelectMenuBuilder}
- */
