@@ -1,64 +1,25 @@
 'use strict';
 
-const Component = require('./Component');
+const process = require('node:process');
+const StringSelectMenuComponent = require('./StringSelectMenuComponent');
+
+let deprecationEmitted = false;
 
 /**
- * Represents a select menu component
- * @extends {Component}
+ * @deprecated Use {@link StringSelectMenuComponent} instead.
+ * @extends {StringSelectMenuComponent}
  */
-class SelectMenuComponent extends Component {
-  /**
-   * The placeholder for this select menu
-   * @type {?string}
-   * @readonly
-   */
-  get placeholder() {
-    return this.data.placeholder ?? null;
-  }
+class SelectMenuComponent extends StringSelectMenuComponent {
+  constructor(...params) {
+    super(...params);
 
-  /**
-   * The maximum amount of options that can be selected
-   * @type {?number}
-   * @readonly
-   */
-  get maxValues() {
-    return this.data.max_values ?? null;
-  }
-
-  /**
-   * The minimum amount of options that must be selected
-   * @type {?number}
-   * @readonly
-   */
-  get minValues() {
-    return this.data.min_values ?? null;
-  }
-
-  /**
-   * The custom id of this select menu
-   * @type {string}
-   * @readonly
-   */
-  get customId() {
-    return this.data.custom_id;
-  }
-
-  /**
-   * Whether this select menu is disabled
-   * @type {?boolean}
-   * @readonly
-   */
-  get disabled() {
-    return this.data.disabled ?? null;
-  }
-
-  /**
-   * The options in this select menu
-   * @type {APISelectMenuOption[]}
-   * @readonly
-   */
-  get options() {
-    return this.data.options;
+    if (!deprecationEmitted) {
+      process.emitWarning(
+        'The SelectMenuComponent class is deprecated, use StringSelectMenuComponent instead.',
+        'DeprecationWarning',
+      );
+      deprecationEmitted = true;
+    }
   }
 }
 
