@@ -12,21 +12,23 @@ interface LabelerData {
 }
 
 export async function createPackage(packageName: string, packageDescription?: string) {
+	const packageDir = join('packages', packageName);
+
 	// Make directory for package
-	await mkdir(`packages/${packageName}`);
+	await mkdir(packageDir);
 
 	// Change to subdirectory
-	chdir(`packages/${packageName}`);
+	chdir(packageDir);
 
 	// Create folder structure
 	await Promise.all([mkdir('src'), mkdir('__tests__')]);
 
+	const templateDir = join('..', 'scripts', 'src', 'template');
+
 	// Create files
-	await writeFile('src/index.ts', `console.log('Hello, from @discord.js/${packageName}');`);
-
-	await writeFile('.eslintrc.json', await readFile('../scripts/src/template/template.eslintrc.json', 'utf8'));
-
-	await writeFile('.lintstagedrc.js', await readFile('../scripts/src/template/template.lintstagedrc.js', 'utf8'));
+	await writeFile(join('src', 'index.ts'), `console.log('Hello, from @discordjs/${packageName}');`);
+	await writeFile('.eslintrc.json', await readFile(join(templateDir, 'template.eslintrc.json'), 'utf8'));
+	await writeFile('.lintstagedrc.js', await readFile(join(templateDir, 'template.lintstagedrc.js'), 'utf8'));
 
 	const packageJSON = {
 		...templateJSON,
@@ -41,10 +43,7 @@ export async function createPackage(packageName: string, packageDescription?: st
 	await writeFile(`package.json`, JSON.stringify(packageJSON, null, 2));
 
 	// Update cliff.toml
-	const cliffTOML = (await readFile(join('..', 'scripts/src/template/cliff.toml'), 'utf8')).replace(
-		'{name}',
-		packageName,
-	);
+	const cliffTOML = (await readFile(join(templateDir, 'cliff.toml'), 'utf8')).replace('{name}', packageName);
 
 	await writeFile('cliff.toml', cliffTOML);
 
@@ -54,21 +53,28 @@ export async function createPackage(packageName: string, packageDescription?: st
 	await writeFile('.cliff-jumperrc.json', JSON.stringify(newCliffJumperJSON, null, 2));
 
 	// Move to github directory
-	chdir('../../.github');
+	chdir(join('..', '..', '.github'));
 
 	const labelsYAML = parseYAML(await readFile('labels.yml', 'utf8')) as LabelerData[];
 	labelsYAML.push({ name: `packages:${packageName}`, color: 'fbca04' });
 
+	labelsYAML.sort((a, b) => a.name.localeCompare(b.name));
+
 	await writeFile('labels.yml', stringifyYAML(labelsYAML));
 
 	const labelerYAML = parseYAML(await readFile('labeler.yml', 'utf8')) as Record<string, string[]>;
-	labelerYAML[`packages/${packageName}`] = [`packages:${packageName}/*`, `packages:${packageName}/**/*`];
+	labelerYAML[`packages:${packageName}`] = [`packages/${packageName}/*`, `packages/${packageName}/**/*`];
 
-	await writeFile('labeler.yml', stringifyYAML(labelerYAML));
+	const sortedLabelerYAML: Record<string, string[]> = {};
+	for (const key of Object.keys(labelerYAML).sort((a, b) => a.localeCompare(b))) {
+		sortedLabelerYAML[key] = labelerYAML[key]!;
+	}
+
+	await writeFile('labeler.yml', stringifyYAML(sortedLabelerYAML));
 
 	// Move back to root
 	chdir('..');
 
 	// Copy default files over
-	await copy('packages/scripts/src/template/default', `packages/${packageName}`);
+	await copy(join('packages', 'scripts', 'src', 'template', 'default'), packageDir);
 }
