@@ -28,7 +28,18 @@ export interface MemberJSON {
 	summary: string | null;
 }
 
-export const PACKAGES = ['builders', 'collection', 'proxy', 'rest', 'util', 'voice', 'ws'];
+export const PACKAGES = [
+	'brokers',
+	'builders',
+	'collection',
+	'core',
+	'formatters',
+	'proxy',
+	'rest',
+	'util',
+	'voice',
+	'ws',
+];
 let idx = 0;
 
 export function createApiModel(data: any) {
