@@ -1,5 +1,7 @@
+import { VscFlame } from '@react-icons/all-files/vsc/VscFlame';
+import { VscInfo } from '@react-icons/all-files/vsc/VscInfo';
+import { VscWarning } from '@react-icons/all-files/vsc/VscWarning';
 import type { PropsWithChildren } from 'react';
-import { VscFlame, VscInfo, VscWarning } from 'react-icons/vsc';
 
 export interface IAlert {
 	title?: string | undefined;
