@@ -1,6 +1,6 @@
+import { VscChevronDown } from '@react-icons/all-files/vsc/VscChevronDown';
 import { Disclosure, DisclosureContent, useDisclosureState } from 'ariakit/disclosure';
 import type { PropsWithChildren } from 'react';
-import { VscChevronDown } from 'react-icons/vsc';
 
 export interface SectionOptions {
 	background?: boolean | undefined;
