@@ -1,8 +1,28 @@
+import { IdentifyThrottler } from '../../utils/IdentifyThrottler.js';
 import type { SessionInfo, WebSocketManager } from '../../ws/WebSocketManager.js';
 import type { FetchingStrategyOptions, IContextFetchingStrategy } from './IContextFetchingStrategy.js';
 
 export class SimpleContextFetchingStrategy implements IContextFetchingStrategy {
-	public constructor(private readonly manager: WebSocketManager, public readonly options: FetchingStrategyOptions) {}
+	// This strategy assumes every shard is running under the same process - therefore we need a single
+	// IdentifyThrottler per manager.
+	private static throttlerCache = new WeakMap<WebSocketManager, IdentifyThrottler>();
+
+	private static ensureThrottler(manager: WebSocketManager): IdentifyThrottler {
+		const existing = SimpleContextFetchingStrategy.throttlerCache.get(manager);
+		if (existing) {
+			return existing;
+		}
+
+		const throttler = new IdentifyThrottler(manager);
+		SimpleContextFetchingStrategy.throttlerCache.set(manager, throttler);
+		return throttler;
+	}
+
+	private readonly throttler: IdentifyThrottler;
+
+	public constructor(private readonly manager: WebSocketManager, public readonly options: FetchingStrategyOptions) {
+		this.throttler = SimpleContextFetchingStrategy.ensureThrottler(manager);
+	}
 
 	public async retrieveSessionInfo(shardId: number): Promise<SessionInfo | null> {
 		return this.manager.options.retrieveSessionInfo(shardId);
@@ -10,5 +30,9 @@ export class SimpleContextFetchingStrategy implements IContextFetchingStrategy {
 
 	public updateSessionInfo(shardId: number, sessionInfo: SessionInfo | null) {
 		return this.manager.options.updateSessionInfo(shardId, sessionInfo);
+	}
+
+	public async waitForIdentify(): Promise<void> {
+		await this.throttler.waitForIdentify();
 	}
 }
